Add initial documentation and configuration files
This commit introduces the initial version of the documentation, including README.adoc, CHANGELOG.md, and associated image files. Updates to the build configuration are also made, adjusting group settings and repository URLs and adding a new Gradle properties file for version management. This setup is part of establishing a MultilingualField add-on for Jmix.